# PERF: avoid extraneous extract_array