--- a/pandas/core/internals/managers.py
+++ b/pandas/core/internals/managers.py
@@ -1809,20 +1809,20 @@ def create_block_manager_from_blocks(
 
 
 def create_block_manager_from_column_arrays(
-    arrays,
+    arrays: list[ArrayLike],
     axes: list[Index],
     consolidate: bool = True,
 ) -> BlockManager:
     # Assertions disabled for performance (caller is responsible for verifying)
     # assert isinstance(axes, list)
     # assert all(isinstance(x, Index) for x in axes)
+    # assert all(isinstance(x, (np.ndarray, ExtensionArray)) for x in arrays)
+    # assert all(type(x) is not PandasArray for x in arrays)
     # assert all(x.ndim == 1 for x in arrays)
     # assert all(len(x) == len(axes[1]) for x in arrays)
     # assert len(arrays) == len(axes[0])
     # These last three are sufficient to allow us to safely pass
     #  verify_integrity=False below.
-
-    arrays = [extract_array(x, extract_numpy=True) for x in arrays]
 
     try:
         blocks = _form_blocks(arrays, consolidate)