```
from asv_bench.benchmarks.frame_ctor import *

self = FromDicts()
self.setup()

%timeit self.time_dict_of_categoricals()
# 530 µs ± 21.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) # <- master
# 471 µs ± 2.85 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each) # <- PR
```